+++ tutorial/TOML/iter_4/toml_demo.rb
@@ -9,25 +9,25 @@ toml_doc = <<-TOML
 
 title = "TOML Example"
 
-[owner]
-name = "Tom Preston-Werner"
-dob = 1979-05-27T07:32:00-08:00
+# [owner]
+# name = "Tom Preston-Werner"
+# dob = 1979-05-27T07:32:00-08:00
 
-[database]
-enabled = true
-ports = [ 8000, 8001, 8002 ]
-data = [ ["delta", "phi"], [3.14] ]
-temp_targets = { cpu = 79.5, case = 72.0 }
+# [database]
+# enabled = true
+# ports = [ 8000, 8001, 8002 ]
+# data = [ ["delta", "phi"], [3.14] ]
+# temp_targets = { cpu = 79.5, case = 72.0 }
 
-[servers]
+# [servers]
 
-[servers.alpha]
-ip = "10.0.0.1"
-role = "frontend"
+# [servers.alpha]
+# ip = "10.0.0.1"
+# role = "frontend"
 
-[servers.beta]
-ip = "10.0.0.2"
-role = "backend"
+# [servers.beta]
+# ip = "10.0.0.2"
+# role = "backend"
 TOML
 
 parser = TOMLParser.new